requirements.txt
Package                      Version
---------------------------- ---------
asgiref                      3.5.0
backports.zoneinfo           0.2.1
certifi                      2021.10.8
charset-normalizer           2.0.12
Django                       4.0.2
django-crispy-forms          1.14.0
django-elasticsearch-dsl     7.2.2
django-elasticsearch-dsl-drf 0.22.4
django-nine                  0.2.5
djangorestframework          3.13.1
elastic-transport            8.0.1
elasticsearch                7.17.0
elasticsearch-dsl            7.4.0
idna                         3.3
pip                          20.1.1
python-dateutil              2.8.2
pytz                         2021.3
requests                     2.27.1
setuptools                   46.4.0
six                          1.16.0
sqlparse                     0.4.2
tzdata                       2021.5
urllib3                      1.26.8
wheel                        0.34.2
